fix: TouchFeedbackViewのHelper追加
